requirements.txt
Flask==2.2
google-cloud-datastore==2.7.1